feat: 優化 check-in 行為, 填完資料直接反應在登記人數
