+++ website/src/components/check-in/check-in.html
@@ -2,7 +2,8 @@
     <link href="/components/check-in/check-in.css" rel="stylesheet">
     <div class="mt-3">
         <div class="form-group position-relative form mb-3">
-            <input data-field="email" type="text" class="form-control" placeholder="輸入信箱搶先登記享終身免費升級進階方案" value="">
+            <input onkeyup="controller.checkInByKeyUp" data-field="email" type="text" class="form-control"
+                placeholder="輸入信箱搶先登記享終身免費升級進階方案" value="">
             <div class="error-msg"></div>
             <button
                 class="btn btn-lg bg-white d-flex align-items-center text-dark-blue mx-auto mt-3 mt-sm-0 hover-bule check-in-btn"
