+++ website/src/components/pricing-plan/pricing-plan.component.js
@@ -1,3 +1,5 @@
+import { RESPONSE_STATUS } from '../../constants.js';
+import { UserDataService } from '../../dataservices/user.dataservice.js';
 import {
     BaseComponent
 } from '../../swim/base.component.js';
@@ -10,7 +12,30 @@ export class PricingPlanComponent extends BaseComponent {
 
     async render () {
         await super.render({
-            ...this.variable
+            ...this.variable,
+            numOfRegisteredUser: 0
         });
+        UserDataService.GetNumberOfRegisteredUsers().then(resp => {
+            if (resp.status === RESPONSE_STATUS.OK) {
+                this.args.numOfRegisteredUser = resp.data.nums;
+            }
+        });
+    }
+
+    async postRender () {
+        await super.postRender();
+        document.body.addEventListener('CONTEXT_NUM_OF_REGISTERED_USER_CHANGED', (event) => {
+            this.variable.numOfRegisteredUser = event.detail.newValue;
+        });
+    }
+
+    computed () {
+        return [{
+            watchKey: 'numOfRegisteredUser',
+            variableName: 'numberOfPromotedUser',
+            value: () => {
+                return this.variable.numOfRegisteredUser && this.variable.numOfRegisteredUser > 300 ? 0 : 300 - this.variable.numOfRegisteredUser;
+            }
+        }];
     }
 }
