+++ website/src/controllers/pricing.controller.js
@@ -1,7 +1,5 @@
 import { APP_CONFIG } from '../config.js';
-import {
-    RoutingController
-} from '../swim/routing-controller.js';
+import { RoutingController } from '../swim/routing-controller.js';
 
 export class PricingController extends RoutingController {
     static get id () {
@@ -16,32 +14,6 @@ export class PricingController extends RoutingController {
             image: `${APP_CONFIG.FRONT_END_URL}/assets/images/share.png`,
             keywords: 'ItemHub,item-hub,物聯網,iot,串聯裝置,連結裝置,早鳥方案,免費升級'
         };
-        await super.render({
-            numOfRegisteredUser: this.args.numOfRegisteredUser
-        });
-    }
-
-    async postRender () {
-        await super.postRender();
-        document.body.addEventListener('CONTEXT_NUM_OF_REGISTERED_USER_CHANGED', this.updateNumOfRegisteredUser);
-    }
-
-    async exit () {
-        document.body.removeEventListener('CONTEXT_NUM_OF_REGISTERED_USER_CHANGED', this.updateNumOfRegisteredUser);
-        return await super.exit();
-    }
-
-    updateNumOfRegisteredUser (event) {
-        this.pageVariable.numOfRegisteredUser = event.detail.newValue;
-    }
-
-    computed () {
-        return [{
-            watchKey: 'numOfRegisteredUser',
-            variableName: 'numberOfPromotedUser',
-            value: () => {
-                return this.pageVariable.numOfRegisteredUser && this.pageVariable.numOfRegisteredUser > 300 ? 0 : 300 - this.pageVariable.numOfRegisteredUser;
-            }
-        }];
+        await super.render({});
     }
 }
