+++ website/src/template/main.html
@@ -23,7 +23,7 @@
         </div>
         <component-feature></component-feature>
         <component-how-to-use></component-how-to-use>
-        <component-pricing-plan bind-number-of-promoted-user="numberOfPromotedUser"></component-pricing-plan>
+        <component-pricing-plan></component-pricing-plan>
         <div class="bottom-check-in">
             <div class="content text-center mx-auto px-3 px-md-0">
                 <h2>收取第一手通知</h2>
