+++ website/src/template/pricing.html
@@ -1,3 +1,3 @@
 <div class="pricing">
-    <component-pricing-plan bind-number-of-promoted-user="numberOfPromotedUser"></component-pricing-plan>
+    <component-pricing-plan></component-pricing-plan>
 </div>
